+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Changes
 
 - New setting to only tilt weapons in third-person when the "Tilt Third Person Items" setting is enabled.
+- Lessen overlay texture allocation when changing damage tint style
 
 ## Fixes